review: clear Aziz policy findings with injectable diagnostics and actor mirror
CmxIrohDebugRelayOverride keeps no new static members; the diag URL is read
through an injectable CmxIrohDebugRelayOverrideDiagnostics struct in its own
package file. The relay diag mirror moves to
MobileHostIrohRuntime+RelayDiag.swift and becomes a revision-ordered actor
instead of an OSAllocatedUnfairLock, read from the iroh_diag Task off-main so
the wedged-main-thread guarantee is unchanged. AppTerminationRequest moves to
its own Sources file.
